+++ R/influxdb_write.R
@@ -8,13 +8,18 @@
 #' In case of a `data.frame`, columns may represent times and both tag and field values.
 #' Columnnames of the `data.frame` object are used as InfluxDB's tag and field keys.
 #' Times and tags are optional. Use parameter `time_col` and `tag_col` to define 
-#' the interpretation. 
+#' the interpretation. By specifiying one of the arguments `measurement` or `measurement_col`, 
+#' a data.frame may contain data from one measurement or multiple measurements, respectively.
 #' 
 #' @inheritParams influx_query
 #' @param x The object to write to an InfluxDB server (either of class `xts` or `data.frame`).
-#' @param measurement Sets the name of the measurement.
+#' @param measurement Sets the name of the measurement (data.frame has data to write
+#' to one measurement only). If both arguments `measurement` and `measurement_col` 
+#' are given, `measurement` gets overridden.
 #' @param time_col A character scalar naming the time index column.
 #' @param tag_cols A character vector naming tag columns.
+#' @param measurement_col A character scalar naming the measurement column (data.frame 
+#' has data to write to mulitple measurements). Overrides `measurement` argument.
 #' @param rp Sets the target retention policy for the write. If not present the
 #' default retention policy is used.
 #' @param precision Sets the precision of the supplied Unix time values
@@ -120,7 +125,7 @@ influx_write.xts <- function(x,
 influx_write.data.frame <- function(x, 
                                     con, 
                                     db, 
-                                    measurement,
+                                    measurement = NULL,
                                     rp = NULL, 
                                     precision = c("s", "ns", "u", "ms", "m", "h"), 
                                     consistency = c(NULL, "one", "quroum", "all", "any"), 
@@ -128,6 +133,7 @@ influx_write.data.frame <- function(x,
                                     use_integers = FALSE,
                                     time_col = NULL, 
                                     tag_cols = NULL,
+                                    measurement_col = NULL,
                                     ...) {
   
   # create query based on function parameters
@@ -158,8 +164,9 @@ influx_write.data.frame <- function(x,
   response <- list_of_df %>% 
     purrr::map( ~ convert_to_line_protocol(x,
                                            measurement = measurement,
-                                           tag_cols = tag_cols,
+                                           measurement_col = measurement_col,
                                            time_col = time_col,
+                                           tag_cols = tag_cols,
                                            precision = precision,
                                            use_integers = use_integers) %>% 
                   httr::POST(body = .,